Tweaked global nav a bit.
Changed to using a {r child=...} snippet for x_web_data_instructions.Rmd -- this has the formatted block telling folks where to go to get web_data, and it loads up the .csv file, too. We may continue to tweak the style and language there, so this has us doing it in just one place and then bubbling up on all pages where it should be included.