Confusion over the "noise" and "intercept"
The noise coefficient is not the same as the intercept.

Confusion is maybe because original was written using y = Ax + e. In this case (i.e. capital A) this would normally be a matrix with x and y being vectors; the intercept falls out of this naturally without needing a separate term.